+++ NEWS.md
@@ -3,6 +3,7 @@
   - `summary.sim`: automatically derives estimate, se, confint parameters if the
      sim routine returns an `estimate` object
   - `merge.estimate`: cast warning if `back.transform` was used
+  - `merge.estimate`: works with objects with and without influence function
   - `summary.sim` df argument for calculating CIs based on t-dist. approximation
 
 # lava 1.8.2